feat(control): add qsh-control + file transfers
- Introduce `qsh-control` crate for shared control protocol and unix socket IPC
- Add `qsh file` subcommands and wire client/server support for file transfer resources
- Refactor qsh-client control layer to re-export protocol/types from `qsh-control`

+++ crates/qsh-client/src/control/commands.rs
@@ -211,7 +211,7 @@ impl FileTransferOptions {
             },
             recursive: self.recursive,
             preserve_mode: false,
-            parallel: self.parallel.max(1),
+            parallel: self.parallel.max(1) as usize,
             skip_if_unchanged: self.skip_unchanged,
         }
     }
